Extracted generic code to utils.
