+++ scripts/utils.py
@@ -1,0 +1,51 @@
+import re
+import shutil
+import tempfile
+from contextlib import contextmanager
+
+
+@contextmanager
+def make_temporary_directory():
+    ''' Context Manager for creating a temporary directory
+    and remove the tree on exit
+
+    Yields
+    ------
+    temp_dir : str
+        absolute path to temporary directory
+    '''
+    try:
+        temp_dir = tempfile.mkdtemp()
+        yield temp_dir
+    finally:
+        shutil.rmtree(temp_dir)
+
+
+def to_camel_case(text, special={'cuds': 'CUDS'}):
+    """ Convert text to CamelCase (for class name)
+
+    Parameters
+    ----------
+    text : str
+        The text to be converted
+
+    special : dict
+        If any substring of text (lower case) matches a key of `special`,
+        the substring is replaced by the value
+
+    Returns
+    -------
+    result : str
+    """
+
+    def replace_func(matched):
+        # word should be lower case already
+        word = matched.group(0).strip("_")
+        if word in special:
+            # Handle special case
+            return special[word]
+        else:
+            # Capitalise the first character
+            return word[0].upper()+word[1:]
+
+    return re.sub(r'(_?[a-zA-Z]+)', replace_func, text.lower())
